requirements.txt
streamlit>=1.35.0
pandas>=2.1.0
boto3
python-dotenv
matplotlib>=3.8.0
numpy>=1.26.0
pillow>=10.0.0
scipy>=1.12.0
